+++ modules/asset/group/src/EventSubscriber/LogEventSubscriber.php
@@ -108,7 +108,7 @@ class LogEventSubscriber implements EventSubscriberInterface {
     }
 
     // If updating an existing group assignment log, invalidate the cache.
-    // This catches group assignment logs changing from done to pending.
+    // This catches group assignment logs changing from done to another status.
     if (!empty($log->original) && $this->isActiveGroupAssignment($log->original)) {
       $update_asset_cache = TRUE;
     }
@@ -155,7 +155,7 @@ class LogEventSubscriber implements EventSubscriberInterface {
     }
 
     // If updating an existing 'done' movement log, invalidate the cache.
-    // This catches any movement logs changing from done to pending.
+    // This catches any movement logs changing from done to another status.
     if (!empty($log->original) && LocationLogEventSubscriber::isActiveMovementLog($log->original)) {
       $update_asset_cache = TRUE;
     }